Replacing roundl by round.

+++ backend/escl/escl.h
@@ -153,7 +153,7 @@ enum
 };
 
 #define PIXEL_TO_MM(pixels, dpi) SANE_FIX((double)pixels * 25.4 / (dpi))
-#define MM_TO_PIXEL(millimeters, dpi) (SANE_Word)roundl(SANE_UNFIX(millimeters) * (dpi) / 25.4)
+#define MM_TO_PIXEL(millimeters, dpi) (SANE_Word)round(SANE_UNFIX(millimeters) * (dpi) / 25.4)
 
 ESCL_Device *escl_devices(SANE_Status *status);
 SANE_Status escl_device_add(int port_nb, const char *model_name,
